+++ Sources/StoreKitPlus/StoreContext+Products.swift
@@ -9,34 +9,49 @@
 import StoreKit
 
 public extension StoreContext {
-    
+
+    /// Check whether a certain product is purchased.
+    func isAnyProductPurchased(
+        in ids: [ProductID]
+    ) -> Bool {
+        ids.contains(where: isProductPurchased)
+    }
+
+    /// Check whether a certain product is purchased.
+    func isAnyProductPurchased(
+        in products: [any ProductRepresentable]
+    ) -> Bool {
+        products.contains(where: isProductPurchased)
+    }
+
+    /// Check whether a certain product is purchased.
+    func isAnyProductPurchased(
+        in product: Product
+    ) -> Bool {
+        products.contains(where: isProductPurchased)
+    }
+
     /// Check whether a certain product is purchased.
     func isProductPurchased(id: ProductID) -> Bool {
         purchasedProductIds.contains(id)
     }
 
     /// Check whether a certain product is purchased.
-    func isProductPurchased(_ product: any ProductRepresentable) -> Bool {
-        isProductPurchased(id: product.id)
+    func isProductPurchased(_ prod: any ProductRepresentable) -> Bool {
+        isProductPurchased(id: prod.id)
     }
 
     /// Check whether a certain product is purchased.
-    func isProductPurchased(_ product: Product) -> Bool {
-        isProductPurchased(id: product.id)
+    func isProductPurchased(_ prod: Product) -> Bool {
+        isProductPurchased(id: prod.id)
     }
 
     /// Get a product from ``products`` with a certain ID.
-    ///
-    /// - Parameters:
-    ///   - id: The ID of the product to fetch.
     func product(withId id: String) -> Product? {
         products.first { $0.id == id }
     }
 
-    /// Get a StoreKit product for a certain product model.
-    ///
-    /// - Parameters:
-    ///   - prod: The local product representation to fetch.
+    /// Get a product from ``products`` for a certain model.
     func product(for prod: any ProductRepresentable) -> Product? {
         product(withId: prod.id)
     }